+++ hosts/zeta/configuration.nix
@@ -1,21 +1,18 @@
-{ config, pkgs, lib, ... }:
+{ ... }:
 
-let
-  inherit (lib) mkIf;
-in {
+{
   imports = [ ./hardware-configuration.nix ];
 
   boot.loader.systemd-boot.enable = true;
   boot.loader.efi.canTouchEfiVariables = true;
 
+  hardware.cpu.amd.updateMicrocode = true;
+
   networking.hostName = "zeta-nixos";
 
-  # Setting `useDHCP` globally is deprecated
-  # manually set `useDHCP` for individual interfaces
-  networking.useDHCP = false;
   networking.interfaces.eno1.useDHCP = true;
 
-  i18n.defaultLocale = "en_US.UTF-8";
+  i18n.defaultLocale = "en_AU.UTF-8";
 
   # GNOME runs Wayland by default
   services.xserver.enable = true;