+++ modules/fonts.nix
@@ -1,0 +1,7 @@
+{
+  nixosModule = { pkgs, ... }: {
+    fonts.fonts = builtins.attrValues {
+      inherit (pkgs) dejavu_fonts;
+    };
+  };
+}