+++ modules/samba.nix
@@ -1,0 +1,14 @@
+{
+  nixosModule = { ... }: {
+    services.samba.enable = true;
+    services.samba.shares.phi = {
+      path = "/";
+      "read only" = "no";
+      "guest ok" = "no";
+      "create mask" = "0644";
+      "directory mask" = "0755";
+      "force user" = "enzime";
+      "force group" = "users";
+    };
+  };
+}